Extract movement into trait
Unfortunately the 'copy' method of case classes is generated by scala,
so it is hard to write a completely generic trait that can copy
itself. However, using F-bounded polymorphism, a trait can at least
define methods that return its subtype. Therefore, logic can be
added in a supertype but the actual copying has to be pushed into the
subtype.